Moved Merge Tag Helpers to helpers.js
… to satisfy ESLint top-level module scope.
